SETR-main/requirements.txt
torch
torchvision
tensorboard
scipy
numpy
matplotlib
opencv_python
tqdm
Pillow
h5py
einops
scikit-learn